% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Print fars data in a plot}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{String containing a number of USA state}

\item{year}{String of a year}
}
\value{
This function plots fars data in specified state
}
\description{
This function gets two strings representing state number and a year
check if state number is valid and plot fars data in map of state
}
\examples{
\dontrun{fars_map_state(52, 2014)}

}
