% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fars data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{String containing the name of the csv file to be read}
}
\value{
This function returns a data structure of type \code{tbl_df}
}
\description{
This function checks if a string passed as argument is a existing csv file,
extract it and returns a \code{tbl_df} data. Treated here as internal function.
}
\note{
If not exists the file an error message is generated
}
