% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read files with fars data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector of string or integer containing years}
}
\value{
This function returns a list of \code{tbl_df} objects
}
\description{
This function gets an array of string or int with years
and returns a list of fars data corresponding to years inputed or
a warning that there is not a fars file in a specific year
}
\note{
It generates a warning if there isn't a file with supplied year in name
}
\examples{
\dontrun{fars_read_years(c(2013, 2014, 2015))}
\dontrun{fars_read_years("2014")}
\dontrun{fars_read_years(1910)}

}
