% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of string or integer containing the years desired to be read}
}
\value{
This function returns a data structure of type \code{tbl_df}
}
\description{
This function gets an array of string or int with years
and returns how many fars data it happened by month of a year
}
\examples{
\dontrun{fars_summarize_years(c(2014, 2015))}

}
