% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generates a filename with a specified year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A vector of numeric integer number}
}
\value{
A filename string with input received
}
\description{
This function takes an object, try to pass it to an integer and return
a string using the received object. Treated here as internal function.
}
\note{
If the vector isn't integer an error is generated
}
